+++ src/main/java/nextstep/courses/domain/enumerate/SessionStatusType.java
@@ -1,4 +1,4 @@
-package nextstep.courses.enumerate;
+package nextstep.courses.domain.enumerate;
 
 public enum SessionStatusType {
     PREPARATION, RECRUITING, END
