+++ src/test/java/nextstep/courses/domain/CourseTest.java
@@ -9,9 +9,9 @@ import java.util.List;
 import nextstep.courses.CanNotJoinException;
 import nextstep.courses.domain.enrollmentcondition.FreeEnrollmentCondition;
 import nextstep.courses.domain.enrollmentcondition.PaidEnrollmentCondition;
-import nextstep.courses.enumerate.CoverImageType;
-import nextstep.courses.enumerate.EnrollmentType;
-import nextstep.courses.enumerate.SessionStatusType;
+import nextstep.courses.domain.enumerate.CoverImageType;
+import nextstep.courses.domain.enumerate.EnrollmentType;
+import nextstep.courses.domain.enumerate.SessionStatusType;
 import nextstep.payments.domain.Payment;
 import nextstep.users.domain.NsUserTest;
 import org.junit.jupiter.api.Test;
