+++ src/test/java/nextstep/courses/domain/CoverImageTest.java
@@ -3,7 +3,7 @@ package nextstep.courses.domain;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
 import nextstep.courses.CanNotCreateException;
-import nextstep.courses.enumerate.CoverImageType;
+import nextstep.courses.domain.enumerate.CoverImageType;
 import org.junit.jupiter.api.Test;
 
 class CoverImageTest {
