+++ src/test/java/nextstep/courses/domain/EnrollmentPolicyTest.java
@@ -6,7 +6,7 @@ import static org.assertj.core.api.Assertions.assertThatThrownBy;
 import nextstep.courses.CanNotJoinException;
 import nextstep.courses.domain.enrollmentcondition.FreeEnrollmentCondition;
 import nextstep.courses.domain.enrollmentcondition.PaidEnrollmentCondition;
-import nextstep.courses.enumerate.SessionStatusType;
+import nextstep.courses.domain.enumerate.SessionStatusType;
 import nextstep.payments.domain.Payment;
 import org.junit.jupiter.api.Test;
